git-svn-id: svn://svn.appwork.org/utils@732 21714237-3853-44ef-a1f0-ef8f03a7d1fe

+++ src/org/appwork/remoteapi/test/TestApiInterface.java
@@ -18,7 +18,7 @@ import org.appwork.remoteapi.RemoteAPIInterface;
 public interface TestApiInterface extends RemoteAPIInterface {
     public String merge(String s1, String s2, int i, boolean b);
 
-    public int sum(int a, int b);
+    public int sum(long a, Byte b);
 
     public boolean toggle(boolean b);
 }
